benchmark/requirements.txt
matplotlib==3.1.1
numpy==1.19.5
pandas==1.1.5
Pillow==6.1.0
scikit-learn==0.23.2
scipy==1.4.1
tensorflow==2.4.1
torch==1.7.0
torchvision==0.8.1
tqdm==4.42.0